Moved diskbuffer to internal since api is not stable
